moved out css to separate files found in src/style
