+++ src/navigation/Routing.js
@@ -12,6 +12,7 @@ import {
   Image
 } from 'react-native';
 import MapScene from '../scenes/MapScene';
+import styles from '../style/routing'
 
 const INDEX = {
   ABOUT: 0,
@@ -120,32 +121,3 @@ export default class Routing extends Component {
 }
 
 
-var styles = StyleSheet.create({
-  app: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'space-between',
-    backgroundColor: '#F5BB94',
-  },
-  Routing: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    height: 80,
-  },
-  NavButton: {
-    padding: 20,
-    flex: 1,
-  },
-  Scene: {
-    flex: 1,
-  },
-  icon: {
-    alignSelf: 'center',
-    width: 27,
-    height: 27,
-  },
-  iconText: {
-    alignSelf: 'center'
-
-  }
-});
