--- src/style/remove.js
@@ -1,1 +1,0 @@
-hej
